refactor(web): share keyboard shortcut specs between modal test suites
